Move stakeholderlayer to layers folder
