+++ src/components/Map.tsx
@@ -4,7 +4,7 @@ import { Stakeholder } from 'types'
 import SearchControl from 'components/controls/SearchControl'
 import InfoPanelControl from 'components/controls/InfoPanelControl'
 import ZoomControl from 'components/controls/ZoomControl'
-import StakeholderLayer from './StakeholderLayer'
+import StakeholderLayer from './layers/StakeholderLayer'
 
 interface MapProps {
   apiKey: string
@@ -16,7 +16,7 @@ const Map: React.FC<MapProps> = ({ apiKey, stakeholders }) => {
   const markersLayer = useRef(null)
 
   return (
-    <MapContainer className="w-full h-full" center={[20, 0]} zoom={3} scrollWheelZoom={true} zoomControl={false}>
+    <MapContainer className="h-full w-full" center={[20, 0]} zoom={3} scrollWheelZoom={true} zoomControl={false}>
       <TileLayer
         attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
         url={`https://tiles.stadiamaps.com/tiles/alidade_smooth/{z}/{x}/{y}{r}.png?api_key=${apiKey}`}
